refactor: BaseEntity 적용 CommenEntity 수정
